identity: move IdentityAllocator providers to identity cell
After removing the `newPolicyTrifecta` function with #34332, the
`IdentityAllocator` is still provided by functions in the daemon package.

This commit moves the function into its own cell in the package
`pkg/identity/cache/cell`. The reason for the subspackage is to break
circular build dependencies.
