+++ daemon/cmd/identity.go
@@ -8,11 +8,8 @@ import (
 
 	"github.com/cilium/cilium/api/v1/models"
 	. "github.com/cilium/cilium/api/v1/server/restapi/policy"
-	"github.com/cilium/cilium/pkg/clustermesh"
 	"github.com/cilium/cilium/pkg/identity"
-	"github.com/cilium/cilium/pkg/identity/cache"
 	identitymodel "github.com/cilium/cilium/pkg/identity/model"
-	"github.com/cilium/cilium/pkg/k8s/client/clientset/versioned"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 )
@@ -63,35 +60,14 @@ func getIdentityEndpointsHandler(d *Daemon, params GetIdentityEndpointsParams) m
 	return NewGetIdentityEndpointsOK().WithPayload(identities)
 }
 
-// CachingIdentityAllocator provides an abstraction over the concrete type in
-// pkg/identity/cache so that the underlying implementation can be mocked out
-// in unit tests.
-type CachingIdentityAllocator interface {
-	cache.IdentityAllocator
-	clustermesh.RemoteIdentityWatcher
-
-	InitIdentityAllocator(versioned.Interface) <-chan struct{}
-
-	// RestoreLocalIdentities reads in the checkpointed local allocator state
-	// from disk and allocates a reference to every previously existing identity.
-	//
-	// Once all identity-allocating objects are synchronized (e.g. network policies,
-	// remote nodes), call ReleaseRestoredIdentities to release the held references.
-	RestoreLocalIdentities() (map[identity.NumericIdentity]*identity.Identity, error)
-
-	// ReleaseRestoredIdentities releases any identities that were restored, reducing their reference
-	// count and cleaning up as necessary.
-	ReleaseRestoredIdentities()
-
-	Close()
-}
-
 func (d *Daemon) AddIdentity(id *identity.Identity) {
 	d.idmgr.Add(id)
 }
+
 func (d *Daemon) RemoveIdentity(id *identity.Identity) {
 	d.idmgr.Remove(id)
 }
+
 func (d *Daemon) RemoveOldAddNewIdentity(old, new *identity.Identity) {
 	d.idmgr.RemoveOldAddNew(old, new)
 }
