+++ daemon/cmd/policy.go
@@ -6,120 +6,29 @@ package cmd
 import (
 	"encoding/json"
 	"fmt"
-	"net"
 	"net/netip"
 	"sync"
 
-	"github.com/cilium/hive/cell"
-	"github.com/cilium/stream"
 	"github.com/go-openapi/runtime/middleware"
 	"github.com/google/uuid"
 
 	"github.com/cilium/cilium/api/v1/models"
 	. "github.com/cilium/cilium/api/v1/server/restapi/policy"
 	"github.com/cilium/cilium/pkg/api"
-	"github.com/cilium/cilium/pkg/clustermesh"
 	"github.com/cilium/cilium/pkg/endpoint"
 	"github.com/cilium/cilium/pkg/endpoint/regeneration"
 	"github.com/cilium/cilium/pkg/eventqueue"
-	"github.com/cilium/cilium/pkg/identity"
-	"github.com/cilium/cilium/pkg/identity/cache"
 	ipcacheTypes "github.com/cilium/cilium/pkg/ipcache/types"
 	"github.com/cilium/cilium/pkg/labels"
 	"github.com/cilium/cilium/pkg/logging/logfields"
 	"github.com/cilium/cilium/pkg/metrics"
 	monitorAPI "github.com/cilium/cilium/pkg/monitor/api"
-	"github.com/cilium/cilium/pkg/node"
-	"github.com/cilium/cilium/pkg/option"
 	"github.com/cilium/cilium/pkg/policy"
 	policyAPI "github.com/cilium/cilium/pkg/policy/api"
 	"github.com/cilium/cilium/pkg/safetime"
 	"github.com/cilium/cilium/pkg/source"
 	"github.com/cilium/cilium/pkg/time"
 )
-
-type identityAllocatorParams struct {
-	cell.In
-
-	Lifecycle        cell.Lifecycle
-	PolicyRepository *policy.Repository
-	PolicyUpdater    *policy.Updater
-}
-
-type identityAllocatorOut struct {
-	cell.Out
-
-	IdentityAllocator      CachingIdentityAllocator
-	CacheIdentityAllocator cache.IdentityAllocator
-	RemoteIdentityWatcher  clustermesh.RemoteIdentityWatcher
-	IdentityObservable     stream.Observable[cache.IdentityChange]
-}
-
-func newIdentityAllocator(params identityAllocatorParams) identityAllocatorOut {
-	// iao: updates SelectorCache and regenerates endpoints when
-	// identity allocation / deallocation has occurred.
-	iao := &identityAllocatorOwner{
-		policy:        params.PolicyRepository,
-		policyUpdater: params.PolicyUpdater,
-	}
-
-	// Allocator: allocates local and cluster-wide security identities.
-	idAlloc := cache.NewCachingIdentityAllocator(iao)
-	idAlloc.EnableCheckpointing()
-
-	params.Lifecycle.Append(cell.Hook{
-		OnStop: func(hc cell.HookContext) error {
-			idAlloc.Close()
-			return nil
-		},
-	})
-
-	return identityAllocatorOut{
-		IdentityAllocator:      idAlloc,
-		CacheIdentityAllocator: idAlloc,
-		RemoteIdentityWatcher:  idAlloc,
-		IdentityObservable:     idAlloc,
-	}
-}
-
-// identityAllocatorOwner is used to break the circular dependency between
-// CachingIdentityAllocator and policy.Repository.
-type identityAllocatorOwner struct {
-	policy        *policy.Repository
-	policyUpdater *policy.Updater
-}
-
-// UpdateIdentities informs the policy package of all identity changes
-// and also triggers policy updates.
-//
-// The caller is responsible for making sure the same identity is not
-// present in both 'added' and 'deleted'.
-func (iao *identityAllocatorOwner) UpdateIdentities(added, deleted identity.IdentityMap) {
-	wg := &sync.WaitGroup{}
-	iao.policy.GetSelectorCache().UpdateIdentities(added, deleted, wg)
-	// Wait for update propagation to endpoints before triggering policy updates
-	wg.Wait()
-	iao.policyUpdater.TriggerPolicyUpdates(false, "one or more identities created or deleted")
-}
-
-// GetNodeSuffix returns the suffix to be appended to kvstore keys of this
-// agent
-func (iao *identityAllocatorOwner) GetNodeSuffix() string {
-	var ip net.IP
-
-	switch {
-	case option.Config.EnableIPv4:
-		ip = node.GetIPv4()
-	case option.Config.EnableIPv6:
-		ip = node.GetIPv6()
-	}
-
-	if ip == nil {
-		log.Fatal("Node IP not available yet")
-	}
-
-	return ip.String()
-}
 
 // PolicyAddEvent is a wrapper around the parameters for policyAdd.
 type PolicyAddEvent struct {
